mikrobot_platform/requirements.txt
# Django Trading Platform Requirements
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7
redis==5.0.1

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.4.0

# Authentication & Security
django-allauth==0.57.0
django-two-factor-auth==1.15.0
cryptography==41.0.7
pyjwt==2.8.0

# Trading Integration
MetaTrader5==5.0.45

# Monitoring & Logging
django-extensions==3.2.3
python-decouple==3.8
sentry-sdk==1.38.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2

# Development
django-debug-toolbar==4.2.0
pytest-django==4.6.0
black==23.9.1
flake8==6.1.0

# Payment Processing (for subscriptions)
stripe==7.8.0

# Real-time Updates
channels==4.0.0
channels-redis==4.1.0
websockets==12.0